refactor: 디렉트리 위치 수정
