+++ src/components/layout/Home/Sections/Sorter.js
@@ -3,7 +3,10 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Link } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import styled from "styled-components";
-import { homeCategoryState, homeSortState } from "../../../recoil/homeRecoil";
+import {
+    homeCategoryState,
+    homeSortState,
+} from "../../../../recoil/homeRecoil";
 
 const StyledSorter = styled.div`
     margin-bottom: 1rem;
